Create deletedomain.py
Created to answer a question asked during Stuart Roger's SAS Viya 3.5 advanced encryption workshop. The question was: While SAS Environment Manager at SAS Viya 3.5 does not allow you to delete an encryption domain (in the Domains page), is it possible to delete an encryption domain from the command line, or using a REST API? We did not find a way to do it using the sas-admin CLI, but I was able to develop this proposed pyviyatool which IS able to delete an encryption domain on my deployment.